Added looping feature

+++ cogs/music.py
@@ -15,7 +15,8 @@ class Music(commands.Cog, name='Music'):
     Can be used by anyone and allows you to listen to music or videos.
     """
     YDL_OPTIONS = {'format': 'bestaudio', 'noplaylist':'True'}
-    FFMPEG_OPTIONS = {'before_options': '-reconnect 1 -reconnect_streamed 1 -reconnect_delay_max 5', 'options': '-vn',}
+    FFMPEG_OPTIONS = {'before_options': '-reconnect 1 -reconnect_streamed 1 -reconnect_delay_max 5', 'options': '-vn'}
+    looping = False
 
     def __init__(self, bot):
         self.bot = bot
@@ -41,6 +42,7 @@ class Music(commands.Cog, name='Music'):
                 .add_field(name='Asked by', value=author)
                 .add_field(name='Uploader', value=f"[{info['uploader']}]({info['channel_url']})")
                 .add_field(name="Queue", value=f"No queued musics")
+                .add_field(name="Looping", value=Music.looping)
                 .set_thumbnail(url=info['thumbnail']))
 
         return {'embed': embed, 'source': info['formats'][0]['url'], 'title': info['title']}
@@ -49,19 +51,27 @@ class Music(commands.Cog, name='Music'):
         embed = self.song_queue[ctx.guild][0]['embed']
         content = "\n".join([f"({self.song_queue[ctx.guild].index(i)}) {i['title']}" for i in self.song_queue[ctx.guild][1:]]) if len(self.song_queue[ctx.guild]) > 1 else "No more songs in queue..."
         embed.set_field_at(index=3, name="Queue:", value=content, inline=False)
+        embed.set_field_at(index=4, name="Looping", value=Music.looping, inline=False)
         await self.message[ctx.guild].edit(embed=embed)
 
-    def play_next(self, ctx):
+    def play_next(self, ctx, looping, cutsong):
         voice = get(self.bot.voice_clients, guild=ctx.guild)
-        if len(self.song_queue[ctx.guild]) > 1:
-            del self.song_queue[ctx.guild][0]
-            run_coroutine_threadsafe(self.edit_message(ctx), self.bot.loop)
-            voice.play(FFmpegPCMAudio(self.song_queue[ctx.guild][0]['source'], **Music.FFMPEG_OPTIONS), after=lambda e: self.play_next(ctx))
+        if not looping:
+            if len(self.song_queue[ctx.guild]) > 1:
+                del self.song_queue[ctx.guild][0]
+                run_coroutine_threadsafe(self.edit_message(ctx), self.bot.loop)
+                voice.play(FFmpegPCMAudio(self.song_queue[ctx.guild][0]['source'], **Music.FFMPEG_OPTIONS), after=lambda e: self.play_next(ctx, Music.looping, cutsong=self.song_queue[ctx.guild][0]['source']))
+                voice.source = PCMVolumeTransformer(voice.source, volume=1.0)
+                voice.is_playing()
+            else:
+                run_coroutine_threadsafe(voice.disconnect(), self.bot.loop)
+                run_coroutine_threadsafe(self.message[ctx.guild].delete(), self.bot.loop)
+        else:
+            voice.play(FFmpegPCMAudio(cutsong, **Music.FFMPEG_OPTIONS), after=lambda e: self.play_next(ctx, Music.looping, cutsong))
             voice.source = PCMVolumeTransformer(voice.source, volume=1.0)
             voice.is_playing()
-        else:
-            run_coroutine_threadsafe(voice.disconnect(), self.bot.loop)
-            run_coroutine_threadsafe(self.message[ctx.guild].delete(), self.bot.loop)
+
+            
 
     @commands.command(aliases=['p'], brief='Du play [url/words]', description='Listen to a video from an url or from a youtube search')
     async def play(self, ctx, *, video: str):
@@ -77,7 +87,6 @@ class Music(commands.Cog, name='Music'):
         await ctx.message.delete()
         
         if not voice.is_playing():
-
 #-------------------------------------------------------------------------------------------------
             async def buttonresume_callback(interaction: discord.Interaction):
                 await interaction.response.send_message("⏯️ Resumed", ephemeral=True, delete_after=10)
@@ -87,38 +96,47 @@ class Music(commands.Cog, name='Music'):
                 await interaction.response.send_message("⏸️ Paused", ephemeral=True, delete_after=10)
                 voice.pause()
 
-            async def buttonqueue_callback(interaction: discord.Interaction):
-                await interaction.response.send_message("🎶 Queued", ephemeral=True, delete_after=10)
-                self.song_queue[ctx.guild].append(song)
-                await self.edit_message(ctx)
+            async def buttonunloop_callback(interaction: discord.Interaction):
+                if not Music.looping:
+                    await interaction.response.send_message("🎶 Looping", ephemeral=True, delete_after=10)
+                    Music.looping = not Music.looping
+                    run_coroutine_threadsafe(self.edit_message(ctx), self.bot.loop)
+                else:
+                    await interaction.response.send_message("🎶 UnLooped", ephemeral=True, delete_after=10)
+                    Music.looping = not Music.looping
+                    run_coroutine_threadsafe(self.edit_message(ctx), self.bot.loop)
             
             async def buttonskip_callback(interaction: discord.Interaction):
-                await interaction.response.send_message("⏭️ Skipped", ephemeral=True, delete_after=10)
-                voice.stop()
+                if not Music.looping:
+                    await interaction.response.send_message("⏭️ Skipped", ephemeral=True, delete_after=10)
+                    voice.stop()
+                else:
+                    await interaction.response.send_message("⏭️ While looping, You can't skip!!", ephemeral=True, delete_after=10)
+
                           
 #-------------------------------------------------------------------------------------------------
-            buttonresume = Button(label="ResumeSong", style=discord.ButtonStyle.blurple, emoji="▶️")
+            buttonresume = Button(label="Resume Song", style=discord.ButtonStyle.blurple, emoji="▶️")
             buttonresume.callback = buttonresume_callback
 
-            buttonpause = Button(label="PauseSong", style=discord.ButtonStyle.red, emoji="⏸️")
+            buttonpause = Button(label="Pause Song", style=discord.ButtonStyle.red, emoji="⏸️")
             buttonpause.callback = buttonpause_callback
 
-            buttonqueue = Button(label="QueueSong", style=discord.ButtonStyle.grey, emoji="🎶")
-            buttonqueue.callback = buttonqueue_callback
+            buttonunloop = Button(label="Loop Song", style=discord.ButtonStyle.grey, emoji="🎶")
+            buttonunloop.callback = buttonunloop_callback
             
-            buttonskip = Button(label="SkipSong", style=discord.ButtonStyle.green, emoji="⏭️")
+            buttonskip = Button(label="Skip Song", style=discord.ButtonStyle.green, emoji="⏭️")
             buttonskip.callback = buttonskip_callback
 
             view = View(timeout=None)
             view.add_item(buttonresume)
             view.add_item(buttonpause)
-            view.add_item(buttonqueue)
+            view.add_item(buttonunloop)
             view.add_item(buttonskip)
 #-------------------------------------------------------------------------------------------------
 
             self.song_queue[ctx.guild] = [song]
             self.message[ctx.guild] = await ctx.send(embed=song['embed'], view=view)
-            voice.play(FFmpegPCMAudio(song['source'], **Music.FFMPEG_OPTIONS), after=lambda e: self.play_next(ctx))
+            voice.play(FFmpegPCMAudio(song['source'], **Music.FFMPEG_OPTIONS), after=lambda e: self.play_next(ctx, Music.looping, song['source']))
             voice.source = PCMVolumeTransformer(voice.source, volume=1.0)
             voice.is_playing()
         else:
@@ -132,7 +150,6 @@ class Music(commands.Cog, name='Music'):
         new_volume = volume/100                   
         voice.source.volume = new_volume
         await interaction.response.send_message(f"Volume: {volume}", ephemeral=True, delete_after=10)
-
 
 async def setup(bot):
     await bot.add_cog(Music(bot))
